pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "dotenv>=0.9.9",
    "ipython>=9.0.2",
    "isort>=6.0.1",
    "langchain>=0.3.20",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.11",
    "langmem>=0.0.15",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "prompts>=0.0.1",
    "pydantic>=2.10.6",
    "typing-extensions>=4.12.2",
]
